+++ eztables/urls.py
@@ -12,6 +12,7 @@ from eztables.demo.views import (
     FormattedBrowserDatatablesView,
     ObjectBrowserDatatablesView,
     CustomBrowserDatatablesView,
+    SpecialCasesDatatablesView,
 )
 
 # Uncomment the next two lines to enable the admin: